Inline iteration methods
This project requires Go 1.25, so build tags are no longer necessary.
